+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [4.7.5] - 2025-XX-XX
 - Update SegmentControl padding
 - Update radio's padding
+- Suffix colors variables for themes
 
 ## [4.7.4] - 2025-08-22
 - Fix iconNameRight Icon (native)