+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [4.8.2] - 2025-09-xx
 - Prevent click on button link disabled
 - Fix Popover web trigger
+- Fix Price striked color
 
 ## [4.8.1] - 2025-09-02
 - Exports enums from Stack