+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Prevent click on button link disabled
 - Fix Popover web trigger
 - Fix Price striked color
+- Fix spinner
 
 ## [4.8.1] - 2025-09-02
 - Exports enums from Stack