+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.8.2] - 2025-09-xx
+## [4.8.2] - 2025-09-17
 - Prevent click on button link disabled
 - Fix Popover web trigger
 - Fix Price striked color